Disable compilation of bindings with test feature
The bindings module does not play nicely with
cargo test. Fortunatley, it's functionality
is tested with pytest, and the module is not
required when testing the core module.

The test suite can now be ran with:
`cargo test --features "test"`

Workaround for: https://github.com/PyO3/pyo3/issues/340 which allows for doctests.
Source of workaround: https://github.com/rust-lang/rust/issues/45599